Rename "src/mapbox-gl.js" to "src/index.js"
