+++ bench/lib/create_map.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const util = require('../../src/util/util');
-const mapboxgl = require('../../src/mapbox-gl');
+const mapboxgl = require('../../src');
 
 module.exports = function createMap(options) {
     options = util.extend({width: 512, height: 512}, options);
